Clarified multiple subnets association purpose
As of now, its not clear why would you want multiple subnets assigned to endpoint. One does not need to associate all subnets in VPC to have access to them - just one is enough. Multiple subnets are assigned just for high availability so if one of the zones to go down, vpn will be still active.